requirements.txt
fastapi>=0.110.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
aiosmtplib>=3.0.0
resend>=2.0.0
pydantic[email]>=2.5.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
mcp>=1.0.0
openai>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0,<1.0
httpx>=0.27.0
